requirements.txt
aiohttp
aiortc
mss
pillow
google-generativeai
python-dotenv
loguru
flask